+++ methods/GSLoopbackBackend/src/constants/actionTypeConstants/index.js
@@ -1,28 +1,28 @@
-// example
-// export const REQUEST_POSTS = 'REQUEST_POSTS'
-// export const RECEIVE_POSTS = 'RECEIVE_POSTS'
-// export const SELECT_SUBREDDIT = 'SELECT_SUBREDDIT'
-// export const INVALIDATE_SUBREDDIT = 'INVALIDATE_SUBREDDIT'
+// // example
+// // export const REQUEST_POSTS = 'REQUEST_POSTS'
+// // export const RECEIVE_POSTS = 'RECEIVE_POSTS'
+// // export const SELECT_SUBREDDIT = 'SELECT_SUBREDDIT'
+// // export const INVALIDATE_SUBREDDIT = 'INVALIDATE_SUBREDDIT'
+// //
 //
-
-export const selectSubreddit = subreddit => ({
-  type: SELECT_SUBREDDIT,
-  subreddit
-})
-
-export const invalidateSubreddit = subreddit => ({
-  type: INVALIDATE_SUBREDDIT,
-  subreddit
-})
-
-export const requestPosts = subreddit => ({
-  type: REQUEST_POSTS,
-  subreddit
-})
-
-export const receivePosts = (subreddit, json) => ({
-  type: RECEIVE_POSTS,
-  subreddit,
-  posts: json.data.children.map(child => child.data),
-  receivedAt: Date.now()
-})
+// export const selectSubreddit = subreddit => ({
+//   type: SELECT_SUBREDDIT,
+//   subreddit
+// })
+//
+// export const invalidateSubreddit = subreddit => ({
+//   type: INVALIDATE_SUBREDDIT,
+//   subreddit
+// })
+//
+// export const requestPosts = subreddit => ({
+//   type: REQUEST_POSTS,
+//   subreddit
+// })
+//
+// export const receivePosts = (subreddit, json) => ({
+//   type: RECEIVE_POSTS,
+//   subreddit,
+//   posts: json.data.children.map(child => child.data),
+//   receivedAt: Date.now()
+// })
