+++ methods/GSLoopbackBackend/src/files.js
@@ -1,13 +1,13 @@
-import {
-  departments,
-  groceries,
-  ingredients,
-  users
-} from '@groceristar/sd-wrapper'
-
-export {
-  departments,
-  groceries,
-  ingredients,
-  users
-}
+// import {
+//   departments,
+//   groceries,
+//   ingredients,
+//   users
+// } from '@groceristar/sd-wrapper'
+//
+// export {
+//   departments,
+//   groceries,
+//   ingredients,
+//   users
+// }
